+++ content-cache-backends-config/src/state.py
@@ -23,6 +23,8 @@ BACKENDS_CONFIG_NAME = "backends"
 PROTOCOL_CONFIG_NAME = "protocol"
 FAIL_TIMEOUT_CONFIG_NAME = "fail-timeout"
 BACKENDS_PATH_CONFIG_NAME = "backends-path"
+HEALTHCHECK_PATH_CONFIG_NAME = "healthcheck-path"
+HEALTHCHECK_INTERVAL_CONFIG_NAME = "healthcheck-interval"
 PROXY_CACHE_VALID_CONFIG_NAME = "proxy-cache-valid"
 
 
@@ -105,6 +107,8 @@ class Configuration(pydantic.BaseModel):
         protocol: The protocol to request the backends with. Can be http or https.
         fail_timeout: The time to wait before using a backend after failure.
         backends_path: The path to request the backends.
+        healthcheck_path: The path to request the healthcheck endpoint.
+        healthcheck_interval: The time between two helthchecks, in milliseconds.
         proxy_cache_valid: The cache valid duration.
     """
 
@@ -126,6 +130,12 @@ class Configuration(pydantic.BaseModel):
         pydantic.StringConstraints(min_length=1),
         pydantic.AfterValidator(_validate_path_value),
     ]
+    healthcheck_path: typing.Annotated[
+        str,
+        pydantic.StringConstraints(min_length=1),
+        pydantic.AfterValidator(_validate_path_value),
+    ]
+    healthcheck_interval: pydantic.PositiveInt
     proxy_cache_valid: tuple[str, ...]
 
     @pydantic.field_validator("proxy_cache_valid")
@@ -173,6 +183,12 @@ class Configuration(pydantic.BaseModel):
             raise ConfigurationError("Empty backends configuration found")
         fail_timeout = typing.cast(str, charm.config.get(FAIL_TIMEOUT_CONFIG_NAME, "")).strip()
         backends_path = typing.cast(str, charm.config.get(BACKENDS_PATH_CONFIG_NAME, "")).strip()
+        healthcheck_path = typing.cast(
+            str, charm.config.get(HEALTHCHECK_PATH_CONFIG_NAME, "")
+        ).strip()
+        healthcheck_interval = typing.cast(
+            int, charm.config.get(HEALTHCHECK_INTERVAL_CONFIG_NAME, -1)
+        )
         proxy_cache_valid_str = typing.cast(
             str, charm.config.get(PROXY_CACHE_VALID_CONFIG_NAME, "")
         ).strip()
@@ -198,6 +214,8 @@ class Configuration(pydantic.BaseModel):
                 protocol=protocol,  # type: ignore
                 fail_timeout=fail_timeout,
                 backends_path=backends_path,
+                healthcheck_path=healthcheck_path,
+                healthcheck_interval=healthcheck_interval,
                 proxy_cache_valid=proxy_cache_valid,  # type: ignore
             )
         except pydantic.ValidationError as err: