+++ content-cache/tests/unit/test_nginx_manager.py
@@ -78,6 +78,8 @@ def test_update_config_with_valid_config(monkeypatch, patch_nginx_manager: None)
                 protocol="https",
                 fail_timeout="30s",
                 backends_path="/backend",
+                healthcheck_path="/health",
+                healthcheck_interval=2123,
                 proxy_cache_valid=("200 302 30m", "404 1m"),
             )
         }
@@ -94,6 +96,10 @@ def test_update_config_with_valid_config(monkeypatch, patch_nginx_manager: None)
     assert "server_name example.com" in config_file_content
     assert "access_log" in config_file_content
     assert "error_log" in config_file_content
+
+    healthchecks_config_file_content = nginx_manager.NGINX_HEALTHCHECKS_CONF_PATH.read_text()
+    assert "GET /health" in healthchecks_config_file_content
+    assert "interval = 2123" in healthchecks_config_file_content
 
 
 def test_health_check(monkeypatch, patch_nginx_manager: None):